requirement.txt
pymupdf
sentence_transformers
pdfplumber
lancedb
pandas
torch
numpy
scipy
Transformers
groq
python-dotenv
streamlit
huggingface_hub
